requirements.txt
requests>=2.31.0
urllib3>=2.1.0
deepface>=0.0.93
python-dotenv>=1.0.0
